pyproject.toml
[project]
name = "migrate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "mysql-connector-python>=9.5.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer[all]>=0.20.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
hamigrate = "hass_migrate.cli:app"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "sqlalchemy>=2.0.45",
    "testcontainers>=4.13.3",
]
